refactor: move test
